Add spec for BugReporterMetricsProxy.getMetrics
