+++ test/helpers/communication/fake-sync-renderer-communication.js
@@ -1,0 +1,53 @@
+/*
+ * Copyright (C) 2018 The "MysteriumNetwork/mysterium-vpn" Authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+// @flow
+
+import type { SyncRendererCommunication } from '../../../src/app/communication/sync/sync-communication'
+import type { SerializedLogCaches } from '../../../src/app/logging/log-cache-bundle'
+import type { RavenData } from '../../../src/app/bug-reporting/metrics/metrics'
+import { TAGS } from '../../../src/app/bug-reporting/metrics/metrics'
+import type { MetricValueDto } from '../../../src/app/bug-reporting/metrics/metric-communication'
+import type { LogDTO } from '../../../src/app/communication/dto'
+
+class FakeSyncRendererCommunication implements SyncRendererCommunication {
+  mockedSerializedCaches: ?SerializedLogCaches = {
+    backend: { info: 'backend info', error: 'backend error' },
+    mysterium_process: { info: 'mysterium info', error: 'mysterium error' },
+    frontend: { info: 'frontend info', error: 'frontend error' }
+  }
+  mockedMetrics: RavenData = {
+    tags: { [TAGS.CLIENT_RUNNING]: true },
+    extra: {}
+  }
+
+  getSerializedCaches () {
+    return this.mockedSerializedCaches
+  }
+
+  sendMetric (dto: MetricValueDto): void {
+  }
+
+  getMetrics (): RavenData {
+    return this.mockedMetrics
+  }
+
+  sendLog (dto: LogDTO) {
+  }
+}
+
+export default FakeSyncRendererCommunication
